Support links between module names in imports and the files defining the module.
The metadata was already created on the module names in the import statements, but we didn't have the corresponding metadata in the file that defines the module.  To support, 2 way linking, the metadata is anchored on the first export keyword in a module.  The tests added cover exports on variable statements as well as stand alone exports, as well as the case when the module is named after a directory, but defined by an index.ts file.